Fix bugfix note
This fixes a bugfix note that was added because of #1293, which
pointed out that a change in a bugfix [1] broke backwards
compatibility.  The bugfix did not work as expected because a
variable was quoted when it should not have been.  This removes
the quotes around the variable, which brings back the expected
functionality.